requirements.txt
# requirements
# Usage: pip install -r requirements.txt

flask
python-dotenv
Pillow>=9.5.0
facenet_pytorch
ultralytics